Refactor checks into separate files
